Temporary fix for vim/vim#3117
It seems that Vim 8 switched to using Python 3. This caused a warning
message to show when launchng Vim as it uses the `imp` dependency which
has been deprecated in favour of `importlib`. This introduces a
temporary fix where we silently execute Python 3 to get rid of the
warning message until it gets fixed in the source code.