Add installation script
Introduces `install.sh` which can be run in Terminal.app on macOS in
order to setup my local development environment. The script is
interactive and will not install anything unless it is asked to. This is
the first iteration of it so it is fairly simple and could do with some
refactoring here and there.