Replace Syntastic with ALE
After trying Syntatsic and getting frustrated with it being slow, I
discovered ALE (Asynchronous Lint Engine) and it is indeed as fast and
convenient as asynchronous can be.

The behaviour of ALE out-of-the-box seems good enough and I like its
minimalism, so I'll keep it without any custom settings for now.