Load asdf in terminal sessions
I've started using asdf to manage Erlang and Elixir versions, and it
needs to load in every new terminal session to work properly. I like the
idea of one single tool to manage multiple language runtime versions, so
I might very well use it for Ruby and Node in the near future.