Add support for local Vim settings
There are some settings that I can't share accross machines or that are
too sensitive to have in this repository. For those, I use local
setting files that I store in my user's home folder.

This introduces the `.local_vim_settings` file that Vim will load if it
exists. It also renames the other local setting files for consistency.